refactor: 필드를 entityClass로 변경
- InsertQueryBuilder는 생성자 파라미터로 entity 받음
