+++ src/main/java/persistence/sql/QueryBuilder.java
@@ -2,8 +2,6 @@ package persistence.sql;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
@@ -12,18 +10,16 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 public abstract class QueryBuilder {
-    private static final Logger logger = LoggerFactory.getLogger(QueryBuilder.class);
-
     public static final String NOT_ENTITY_FAILED_MESSAGE = "클래스에 @Entity 애노테이션이 존재해지 않습니다.";
 
-    private final Object entity;
+    private final Class<?> entityClass;
 
-    public QueryBuilder(Object entity) {
-        if (!entity.getClass().isAnnotationPresent(Entity.class)) {
+    public QueryBuilder(Class<?> entityClass) {
+        if (!entityClass.isAnnotationPresent(Entity.class)) {
             throw new IllegalArgumentException(NOT_ENTITY_FAILED_MESSAGE);
         }
 
-        this.entity = entity;
+        this.entityClass = entityClass;
     }
 
     public abstract String build();
@@ -33,17 +29,8 @@ public abstract class QueryBuilder {
     }
 
     protected List<Field> getColumns() {
-        return Arrays.stream(entity.getClass().getDeclaredFields())
+        return Arrays.stream(entityClass.getDeclaredFields())
                 .collect(Collectors.toList());
-    }
-
-    protected Object getValue(Field field) {
-        try {
-            return field.get(entity);
-        } catch (IllegalAccessException e) {
-            logger.error(e.getMessage(), e);
-            return null;
-        }
     }
 
     private Object[] getArgs(String[] templateArgs) {
@@ -54,12 +41,11 @@ public abstract class QueryBuilder {
     }
 
     private String getTableName() {
-        final Table table = entity.getClass().getAnnotation(Table.class);
+        final Table table = entityClass.getAnnotation(Table.class);
         if (Objects.nonNull(table) && Objects.nonNull(table.name()) && !table.name().isBlank()) {
             return table.name();
         }
-        return entity.getClass()
-                .getSimpleName()
+        return entityClass.getSimpleName()
                 .toLowerCase();
     }
 }
