+++ src/main/java/persistence/sql/ddl/CreateQueryBuilder.java
@@ -13,8 +13,8 @@ public class CreateQueryBuilder extends QueryBuilder {
     private static final String GENERATION_COLUMN_DEFINITION = "AUTO_INCREMENT";
     private static final String PRIMARY_KEY_COLUMN_DEFINITION = "PRIMARY KEY";
 
-    public CreateQueryBuilder(Object entity) {
-        super(entity);
+    public CreateQueryBuilder(Class<?> entityClass) {
+        super(entityClass);
     }
 
     @Override
