+++ src/main/java/persistence/sql/ddl/DropQueryBuilder.java
@@ -5,8 +5,8 @@ import persistence.sql.QueryBuilder;
 public class DropQueryBuilder extends QueryBuilder {
     private static final String QUERY_TEMPLATE = "DROP TABLE IF EXISTS %s";
 
-    public DropQueryBuilder(Object entity) {
-        super(entity);
+    public DropQueryBuilder(Class<?> entityClass) {
+        super(entityClass);
     }
 
     @Override
