+++ src/test/java/persistence/sql/ddl/CreateQueryBuilderTest.java
@@ -12,11 +12,8 @@ class CreateQueryBuilderTest {
     @Test
     @DisplayName("@Entity 애노테이션이 존재하는 클래스로 인스턴스를 생성한다.")
     void constructor() {
-        // given
-        final Person person = new Person("Jaden", 30, "test@email.com", 1);
-
         // when
-        final CreateQueryBuilder createQueryBuilder = new CreateQueryBuilder(person);
+        final CreateQueryBuilder createQueryBuilder = new CreateQueryBuilder(Person.class);
 
         // then
         assertThat(createQueryBuilder).isNotNull();
@@ -25,9 +22,6 @@ class CreateQueryBuilderTest {
     @Test
     @DisplayName("@Entity 애노테이션이 존재하지 않는 클래스로 인스턴스를 생성하면 예외가 발생한다.")
     void constructor_exception() {
-        // given
-        final NotEntity notEntity = new NotEntity("Jaden", 30, "test@email.com", 1);
-
         // when & then
         assertThatThrownBy(() -> new CreateQueryBuilder(NotEntity.class))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -38,8 +32,7 @@ class CreateQueryBuilderTest {
     @DisplayName("CREATE 쿼리를 생성한다.")
     void build() {
         // given
-        final Person person = new Person("Jaden", 30, "test@email.com", 1);
-        final CreateQueryBuilder createQueryBuilder = new CreateQueryBuilder(person);
+        final CreateQueryBuilder createQueryBuilder = new CreateQueryBuilder(Person.class);
 
         // when
         final String query = createQueryBuilder.build();
