+++ src/test/java/persistence/sql/ddl/DropQueryBuilderTest.java
@@ -12,11 +12,8 @@ class DropQueryBuilderTest {
     @Test
     @DisplayName("@Entity 애노테이션이 존재하는 클래스로 인스턴스를 생성한다.")
     void constructor() {
-        // given
-        final Person person = new Person("Jaden", 30, "test@email.com", 1);
-
         // when
-        final DropQueryBuilder dropQueryBuilder = new DropQueryBuilder(person);
+        final DropQueryBuilder dropQueryBuilder = new DropQueryBuilder(Person.class);
 
         // then
         assertThat(dropQueryBuilder).isNotNull();
@@ -25,11 +22,8 @@ class DropQueryBuilderTest {
     @Test
     @DisplayName("@Entity 애노테이션이 존재하지 않는 클래스로 인스턴스를 생성하면 예외가 발생한다.")
     void constructor_exception() {
-        // given
-        final NotEntity notEntity = new NotEntity("Jaden", 30, "test@email.com", 1);
-
         // when & then
-        assertThatThrownBy(() -> new CreateQueryBuilder(notEntity))
+        assertThatThrownBy(() -> new CreateQueryBuilder(NotEntity.class))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(QueryBuilder.NOT_ENTITY_FAILED_MESSAGE);
     }
@@ -38,8 +32,7 @@ class DropQueryBuilderTest {
     @DisplayName("DROP 쿼리를 생성한다.")
     void build() {
         // given
-        final Person person = new Person("Jaden", 30, "test@email.com", 1);
-        final DropQueryBuilder dropQueryBuilder = new DropQueryBuilder(person);
+        final DropQueryBuilder dropQueryBuilder = new DropQueryBuilder(Person.class);
 
         // when
         final String query = dropQueryBuilder.build();
